Auth revamp and notifications and profilepictures
Auth now runs through provider and uses a firstTimeLogin variable to go to setUsername for all kinds of logins.

You can now set a profilepicture, which uploads to storage and saves the url in the user profile.

Notifications now exist, two kinds are supported for now. They are shown with tiny profile pictures. Noise.